Fix test failures in test_sync_noflush_novote, take 2
The doTicks approach did not work well because it isn't very accurate and doesn't take into account the time spent outside of the actual ticking.
So this replaces it with a direct loop which compares against the current time to ensure that exactly the right amount of time has elapsed since the creation of the SyncObj.

This test may break again if the time spent inside SyncObj increases greatly; for example, if __init__ takes very long, the start time would not match the corresponding values inside the objects. Each individual tick may also not take too long.